requirements.txt
Django
psycopg2-binary
pgvector
sentence-transformers
streamlit
redis